kernel: expose allowlist workaround as Kconfig option + ++ booted
Useful for situations where the SU allowlist is not kept after a reboot.
As per upstream this is only used for < 4.10 and Huawei HiSilicon devices.
but theres user reports having issues even on 4.14/4.19 samsung kernels.
Expose this option so users affected can opt-in.

This supercedes ` kernel:Add Huawei hisi check (tiann#1545) `

References:
	https://github.com/tiann/KernelSU/commit/f57d351adb50663ba3cd3ce8e8c8c615b62cf7de
	https://github.com/tiann/KernelSU/commit/b61cb30c8832b9433d782674a81f7ccf1fd40c27

Reviewed-by: Alex <a.mihail@pm.me>
Signed-off-by: backslashxx <118538522+backslashxx@users.noreply.github.com>
Co-Authored-By: ExtremeXT <75576145+extremext@users.noreply.github.com>